# docs/swagger.yaml
definitions:
...
  models.UserKYC:
    properties:
      created_at:
        type: string
      id:
        type: integer
      id_photo_back:
        type: string
      id_photo_front:
        type: string
      identity_card:
        type: string
      name:
        type: string
      photo:
        type: string
      photo_with_id:
        type: string
      updated_at:
        type: string
      user_id:
        type: integer
    type: object
...
  status.Response:
    properties:
      status_code:
        type: integer
      status_message:
        type: string
    type: object
  status.ResponseWtihData:
    properties:
      data:
        type: any
      status:
        $ref: '#/definitions/status.Response'
    type: object
info:
  contact:
    email: service@invar.finance
    name: InVar
    url: https://www.invar.finance/
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
paths:
...
  /admin/kyc/:id:
    patch:
      consumes:
      - multipart/form-data
      description: 管理者更新實名資料
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      - description: 姓名
        in: formData
        name: name
        type: string
      - description: 身分證號碼
        in: formData
        name: identity_card
        type: string
      - description: 審核結果
        in: formData
        name: status
        type: string
      - description: 大頭貼
        in: formData
        name: photo_file
        type: file
      - description: 身分證正面
        in: formData
        name: id_photo_front_file
        type: file
      - description: 身分證背面
        in: formData
        name: id_photo_back_file
        type: file
      - description: 與身分證合照
        in: formData
        name: photo_with_id_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 管理者更新實名資料
      tags:
      - User
    post:
      consumes:
      - application/json
      description: 管理者獲得使用者實名資料
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.UserKYC'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 管理者獲得使用者實名資料
      tags:
      - User
...
  /kyc:
    get:
      consumes:
      - application/json
      description: 獲得實名資料
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/status.ResponseWtihData'
            - properties:
                data:
                  $ref: '#/definitions/models.UserKYC'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 獲得實名資料
      tags:
      - User
    patch:
      consumes:
      - multipart/form-data
      description: 更新實名資料
      parameters:
      - description: 姓名
        in: formData
        name: name
        type: string
      - description: 身分證號碼
        in: formData
        name: identity_card
        type: string
      - description: 大頭貼
        in: formData
        name: photo_file
        type: file
      - description: 身分證正面
        in: formData
        name: id_photo_front_file
        type: file
      - description: 身分證背面
        in: formData
        name: id_photo_back_file
        type: file
      - description: 與身分證合照
        in: formData
        name: photo_with_id_file
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 更新實名資料
      tags:
      - User
    post:
      consumes:
      - multipart/form-data
      description: 新增實名資料
      parameters:
      - description: 姓名
        in: formData
        name: name
        required: true
        type: string
      - description: 身分證號碼
        in: formData
        name: identity_card
        required: true
        type: string
      - description: 大頭貼
        in: formData
        name: photo_file
        required: true
        type: file
      - description: 身分證正面
        in: formData
        name: id_photo_front_file
        required: true
        type: file
      - description: 身分證背面
        in: formData
        name: id_photo_back_file
        required: true
        type: file
      - description: 與身分證合照
        in: formData
        name: photo_with_id_file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/status.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/status.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/status.Response'
      security:
      - BearerAuth: []
      summary: 新增實名資料
      tags:
      - User
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"